+++ src/internal/HRTFDatabaseLoader.h
@@ -25,14 +25,6 @@ public:
     // @CBB the guts of the loader should be a private singleton, so that the loader can be constructed without a factory
     explicit HRTFDatabaseLoader(float sampleRate, const std::string & searchPath);
 
-    // Lazily creates the singleton HRTFDatabaseLoader (if not already created) and starts loading asynchronously (when created the first time).
-    // Returns the singleton HRTFDatabaseLoader.
-    // Must be called from the main thread.
-    static std::shared_ptr<HRTFDatabaseLoader> MakeHRTFLoaderSingleton(float sampleRate, const std::string & searchPath);
-
-    // Returns the singleton HRTFDatabaseLoader.
-    static std::shared_ptr<HRTFDatabaseLoader> loader() { return s_loader; }
-
     // Both constructor and destructor must be called from the main thread.
     ~HRTFDatabaseLoader();
 
@@ -49,11 +41,6 @@ public:
     // Called in asynchronous loading thread.
     void load();
 
-    // defaultHRTFDatabase() gives access to the loaded database.
-    // This can be called from any thread, but it is the callers responsibilty to call this while the context (and thus HRTFDatabaseLoader)
-    // is still alive.  Otherwise this will return 0.
-    static HRTFDatabase * defaultHRTFDatabase();
-
     // If it hasn't already been loaded, creates a new thread and initiates asynchronous loading of the default database.
     // This must be called from the main thread.
     void loadAsynchronously();
@@ -61,8 +48,6 @@ public:
 private:
     static void databaseLoaderEntry(HRTFDatabaseLoader * threadData);
 
-
-    static std::shared_ptr<HRTFDatabaseLoader> s_loader;  // singleton
     std::unique_ptr<HRTFDatabase> m_hrtfDatabase;
 
     // Holding a m_threadLock is required when accessing m_databaseLoaderThread.